requirements.txt
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
asttokens==2.0.5
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Django==5.0.2
djangorestframework==3.14.0
entrypoints==0.4
executing==0.8.3
idna==3.3
ipykernel==6.9.2
ipython==8.1.1
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.0.3
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyter-server==1.15.6
jupyterlab==3.3.2
jupyterlab-pygments==0.1.2
jupyterlab-server==2.11.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclassic==0.3.7
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
notebook==6.4.10
notebook-shim==0.1.0
numpy==1.22.3
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
pillow==10.2.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
psutil==5.9.0
psycopg2==2.9.9
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
pywin32==303
pywinpty==2.0.5
pyzmq==22.3.0
requests==2.27.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
sqlparse==0.4.4
stack-data==0.2.0
terminado==0.13.3
testpath==0.6.0
tornado==6.1
traitlets==5.1.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.9
vboxapi==1.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.1
